buildings
the 00_CastleCulture file in the old mod was mostly vanilla so I
stripped out the custom bit and put it in 01_CastleCulture. The order
house in special buildings is supposed to only be buildable by holy
orders. Not sure if that syntax will work, however. May need to manually
put in "has_landed_title = " for all of the holy orders. Which will be a
pain.