Fix for bug re: nonstandard whitespace characters
(thanks to CyberSpyder):

"The space after the = in the working event is a normal space, ascii hex
code 20. The space after the = in the broken one is ascii hex code A0, a
non-breaking space which I believe is interpreted by the game as part of
the border name. Since the border " GFX_event_normal_frame_intrigue"
doesn't exist, it avoids firing the event...though apparently it will
still force something through if you trigger it with the console."