Fix to "restoring religious head" decisions
Religious head will now be chosen from the priests holding holy sites
within the restorer's realm (so they will have land and can declare GHWs
if available).

If no priests hold holy sites in the restorer's realm (how the hell did
they get 60% MA then?) a random priest in the restorer's realm will be
promoted.

(Pagan reformation automatically assigns land and Restore the Pontifex
gives the religious head title to the restorer, so this is not necessary
for those faiths.