removed redundant set_global_flag
the event called right after that flag is set sets that flag. Added an
if statement to LoR.21 that checks for ROOT already holding the
e_roman_empire title so it'll skip the title granting and law copying
step if the event was called via the title being created regularly.