simplified job spiritual localization
All pagan lord spirituals are "Seer"s or "Seeress"es unless they're a
king of pagan with a specific localization for it. Religious lord
spiritual titles now carry across government types. Court Chaplains are
now only Chaplains if they're in a government type with no Court (e.g.
Nomads and Tribes). Didn't adjust the Imams or Mullahs because I don't
know the languages well enough to pull  "Court" out of them so Nomadic
and Tribal Imams and Mullahs are still Imams and Mullahs.