Adjustments
Removed Ard Ri as it is only really appropriate for Ireland which has
custom localization on the title anyway. Ri only really applies to the
Irish as the Scottish have a different word for it and the Pictish
didn't speak Gaelic. Removed the "Llwyth of" localization as it's only
welsh and the welsh aren't tribal at any time in-game. Removed Riocht
localization because it literally just means "Kingdom" in Gaelic.