Minor name alterations to old Saxon and the Elbers
made a few name alterations, namely removed Hedwig from old saxon, and
replaced it with perhaps a low german alternative spelling to replace
it. the name Elisabeth is no longer available to the more isolated
germans (like Swiss and Thuringian and old German) but is present in
those close to latin and low german influences. not much changed beyond
that.