Galatian and Pontic salvaged
I had the time, so hear is the extra icing on the cake. Galatian and
Pontic are done. since they will only be existing as local cultures that
are meant to increase difficulty of managing the empire there is no need
to do any further work unless we feel otherwise needed later on.