TWEAK: Can't load non-plus saves
Added user_dir to mod files. This will make it so users can not load
saves started without plus. Also added a SECOND user_dir to the india
file. Through experimentation, discovered that it creates another folder
so that an India save can't be loaded using just the main mod. Did it
all in one commit for easier reversion if it is not a change wanted.