Moved Norse Holy Site back to Norduland Eystra
Meant to represent the Waterfall of the Gods. Increased settlement limit on both Norduland provinces to 3 to make sure a temple can always be built there and maintain the balance of Iceland's split. Added history of temple being built in 1000 to represent the holy site, added temple to landed_titles due to none of the existing options were appropriate for the location.