History Correction
ERE does not exist prior to 330CE. ERE title should start as "Eastern Roman Empire" and switch to "Byzantine Empire" at the coronation of Charlemagne or Holy Roman Emperor. History "name =" and "adjective =" commands don't seem to be working on e_byzantium for some reason, so used effects in history instead.

(I will not be upset if this is reverted. That's why I did it as one commit)